feat(search_service): update search service to use the new SearchHighlight
